Vehicle and Dof refactored more
I've refactored Vehicle and Dof classes to make all member
variables private and to include getter and setter functions.
This, I think, is as much as we'll be able to do without testing
the two classes. So, now we need to develop an automated test
suite for testing the functionality of these two classes thus
far (yes, I know Vehicle is *incomplete* since the EKF is not
written nor integrated into Vehicle yet).