Updated EKF and Finally Built Code
Updated EKF to take in the mass of the vehicle as an argument to functions that
need it rather than relying on a macro constant for it. Also, slightly refactored
ekf code to make it more readable and consistent with the style throughout the
rest of the controls code.

Finally figured out that even though CCS's code anaylzer tool may not always find
standard library functions (i.e. memcpy), the code still builds. So, I've turned
off the analysis launch so that it doesn't give us errors. The code builds
just fine and a recent experiment proved that memcpy still works.