Huge Changes to EKF and Runnables
Updated EKF to reflect the new derivation of it in the global frame only and
only for the 3 inertial states and their derivatives (so 6 only). Also
included the passing in of Euler Angles as an argument becuase they are
part of the coeficients of the state equations and not members of the state
vectors themselves in our EKF. All relevant files were updated (ekf files that
is). Also updated CtrlRunnable and Vehicle to include these changes. Not only
that, we don't need Vehicle's default ctor and the vlog values for the euler
angles will go to 0 until further notice. The i2c defines of this particular
project have been updated to reflect the address of the px4 currently on the
fully built vehicle we're testing with. CtrlRunnable also now handles the
different control modes and the two different ekf sensor updates based on when a
new setpt message comes in and what the flight mode is. Dji Runnable has a
temporary safety feature that passes the manual control through to the states
of the DJI that we aren't controlling while in the tuning process. Only the
state PID(s) we want to tune are passed into the DJI as custom signals.