EKF Timing and Model Fixes; New Logs and Parsing
Fixed derivation issues in f(x,u) in ekf model as well as the timing of the
prediciton and correction steps. f(x,u) now doesn't have the addition of gravity
to its az component since the control input of Body-Frame Fz to the DJI already
has the mg-offset for the vehicle. Also fixed the timing of the correction step
so that it only occurs every 20 ms (rate of Lidar updates, which are the slowest
of all sensors) while prediction happens in between at every 10 ms (timing of
the control runnable). Code builds but needs to be flight-tested.

Also updated spKalman script to plot better. It now shows the filtered state
at each iteration with covaraince boundaries in a format that we can use for
tuning. Now the ekf in the code on the Tiva matches the ekf in the matlab
script, which is what we want since the matlab one we created was correct.
Corresponding log files have also been added.